Find simple CpG (CimpleG) signatures.

Train a classification model using (CpGs) as features for the given target data.

CimpleG(
  train_data,
  train_targets = NULL,
  targets,
  test_data = NULL,
  test_targets = NULL,
  method = c("CimpleG", "CimpleG_parab", "CimpleG_unscaled", "brute_force", "oner",
    "logistic_reg", "decision_tree", "boost_tree", "mlp", "rand_forest", "null_model"),
  pred_type = c("both", "hypo", "hyper"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  k_folds = 10,
  grid_n = 10,
  param_p = 4,
  quantile_threshold = 0.01,
  train_only = FALSE,
  run_parallel = FALSE,
  save_dir = NULL,
  save_format = c("zstd", "lz4", "gzip", "bzip2", "xz", "nocomp"),
  verbose = 1
)

Arguments

train_data

Training dataset. A matrix (s x f) with methylation data (Beta values) that will be used to train/find the predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns.

train_targets

A data frame with the training target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `train_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`.

targets

A string specifying the name of the column in `train_targets` to be used for training. Can be a character vector if there are several columns in `train_targets` to be used for training. If this argument is a character vector, CimpleG will search for the best predictors for each target sequentially or in parallel depending on the value of `run_parallel`

test_data

Testing dataset. A matrix (s x f) with methylation data (Beta values) that will be used to test the performance of the found predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

test_targets

A data frame with the testing target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `test_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

method

A string specifying the method or type of machine learning model/algorithm to be used for training. These are divided in two main groups. * The simple models (classifiers that use a single feature), `CimpleG` (default), `brute_force`, `CimpleG_unscaled` or `oner`; * the complex models (classifiers that use several features), `logistic_reg`, `decision_tree`, `boost_tree`, `mlp` or `rand_forest`.

pred_type

A string specifying the type of predictor/CpG to be searched for during training. Only used for simple models. One of `both` (default), `hypo` or `hyper`. If `hypo`, only hypomethylated predictors will be considered. If `hyper`, only hypermethylated predictors will be considered.

engine

A string specifying the machine learning engine behind `method`. Only used for complex models. Currently not in use.

k_folds

An integer specifying the number of folds (K) to be used in training for the stratified K-fold cross-validation procedure.

grid_n

An integer specifying the number of hyperparameter combinations to train for.

param_p

An even number in `sigma / (delta^param_p)`. Tunes how much weight will be given to delta when doing feature selection. Default is 4.

quantile_threshold

A number between 0 and 1. Determines how many features will be kept. Default is 0.001.

train_only

A boolean, if TRUE, CimpleG will only train (find predictors) but not test them against a test dataset.

run_parallel

A boolean, if `FALSE`, the default, it will search for predictors for multiple targets sequentially. If `TRUE` it will search for predictors for multiple targets at the same time (parallel processing) in order to save in computational time. You need to set up `future::plan()` before running this function.

save_dir

If defined it will save the resulting model to the given directory. Default is NULL.

save_format

Only used if save_dir is not NULL. One of "zstd", "lz4", "gzip", "bzip2","xz", "nocomp". zstd is the best option, fast compression and loading times, low space usage.

verbose

How verbose you want CimpleG to be while it is running. At 0, no message is displayed, at 3 every message is displayed. Default is 1.

Value

A CimpleG object with the results per target class.

Examples

library("CimpleG")

# read data
data(train_data)
data(train_targets)
data(test_data)
data(test_targets)

# run CimpleG
cimpleg_result <- CimpleG(
  train_data = train_data,
  train_targets = train_targets,
  test_data = test_data,
  test_targets = test_targets,
  method = "CimpleG",
  targets = c("CELL_TYPE_MSCORFIBRO","CELL_TYPE_NEURONS")
)
#> Training for target 'CELL_TYPE_MSCORFIBRO' with 'CimpleG' has finished.: 4.94 sec elapsed
#> Training for target 'CELL_TYPE_NEURONS' with 'CimpleG' has finished.: 0.643 sec elapsed

# check results
cimpleg_result$results
#> $CELL_TYPE_MSCORFIBRO
#> $CELL_TYPE_MSCORFIBRO$train_res
#> $CELL_TYPE_MSCORFIBRO$train_res$fold_id
#> # A tibble: 4,090 × 3
#>      Row Data     Fold  
#>    <int> <chr>    <chr> 
#>  1     1 Analysis Fold02
#>  2     1 Analysis Fold03
#>  3     1 Analysis Fold04
#>  4     1 Analysis Fold05
#>  5     1 Analysis Fold06
#>  6     1 Analysis Fold07
#>  7     1 Analysis Fold08
#>  8     1 Analysis Fold09
#>  9     1 Analysis Fold10
#> 10     2 Analysis Fold01
#> # … with 4,080 more rows
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_summary
#>             id     stat_origin mean_aupr fold_presence
#>  1: cg02023138      train_aupr 0.8326981             6
#>  2: cg02023138 validation_aupr 0.8097305             6
#>  3: cg03369247      train_aupr 0.9308328            10
#>  4: cg03369247 validation_aupr 0.9306799            10
#>  5: cg04232282      train_aupr 0.7348655            10
#>  6: cg04232282 validation_aupr 0.7717972            10
#>  7: cg06932535      train_aupr 0.6618693            10
#>  8: cg06932535 validation_aupr 0.6990152            10
#>  9: cg11029367      train_aupr 0.7914593            10
#> 10: cg11029367 validation_aupr 0.8078551            10
#> 11: cg14021880      train_aupr 0.9485523            10
#> 12: cg14021880 validation_aupr 0.9516236            10
#> 13: cg15324723      train_aupr 0.8774878            10
#> 14: cg15324723 validation_aupr 0.8772958            10
#> 15: cg16427420      train_aupr 0.6861396            10
#> 16: cg16427420 validation_aupr 0.7366863            10
#> 17: cg17786959      train_aupr 0.8718700            10
#> 18: cg17786959 validation_aupr 0.8962509            10
#> 19: cg24192660      train_aupr 0.9015768            10
#> 20: cg24192660 validation_aupr 0.8981027            10
#> 21: cg26461417      train_aupr 0.9371736             4
#> 22: cg26461417 validation_aupr 0.8862681             4
#>             id     stat_origin mean_aupr fold_presence
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$dt_dmsv
#>             id diff_means sum_variance pred_type
#>  1: cg02023138 -0.5213741   0.07679563     FALSE
#>  2: cg03369247 -0.6070865   0.06291659     FALSE
#>  3: cg04232282 -0.5580492   0.09075907     FALSE
#>  4: cg06932535 -0.5998086   0.07776802     FALSE
#>  5: cg11029367 -0.6247012   0.09282683     FALSE
#>  6: cg14021880 -0.5730696   0.06488891     FALSE
#>  7: cg15324723 -0.5215541   0.06901091     FALSE
#>  8: cg16427420 -0.5869993   0.07927203     FALSE
#>  9: cg17786959 -0.4598171   0.04412714     FALSE
#> 10: cg24192660  0.6911130   0.13719117      TRUE
#> 11: cg26461417  0.4642382   0.04979224      TRUE
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_results
#>             id     stat_origin mean_aupr fold_presence diff_means sum_variance
#>  1: cg24192660 validation_aupr 0.8981027            10  0.6911130   0.13719117
#>  2: cg03369247 validation_aupr 0.9306799            10 -0.6070865   0.06291659
#>  3: cg11029367 validation_aupr 0.8078551            10 -0.6247012   0.09282683
#>  4: cg14021880 validation_aupr 0.9516236            10 -0.5730696   0.06488891
#>  5: cg06932535 validation_aupr 0.6990152            10 -0.5998086   0.07776802
#>  6: cg16427420 validation_aupr 0.7366863            10 -0.5869993   0.07927203
#>  7: cg04232282 validation_aupr 0.7717972            10 -0.5580492   0.09075907
#>  8: cg15324723 validation_aupr 0.8772958            10 -0.5215541   0.06901091
#>  9: cg02023138 validation_aupr 0.8097305             6 -0.5213741   0.07679563
#> 10: cg17786959 validation_aupr 0.8962509            10 -0.4598171   0.04412714
#> 11: cg26461417 validation_aupr 0.8862681             4  0.4642382   0.04979224
#>     pred_type    cpg_score train_rank
#>  1:      TRUE 8.981027e-01          1
#>  2:     FALSE 5.926112e-01          2
#>  3:     FALSE 5.759198e-01          3
#>  4:     FALSE 4.660049e-01          4
#>  5:     FALSE 4.231055e-01          5
#>  6:     FALSE 4.051134e-01          6
#>  7:     FALSE 3.278291e-01          7
#>  8:     FALSE 2.342303e-01          8
#>  9:     FALSE 7.186803e-02          9
#> 10:     FALSE 8.962509e-05         10
#> 11:      TRUE 1.702745e-06         11
#> 
#> 
#> $CELL_TYPE_MSCORFIBRO$test_perf
#>             id validation_mean_aupr fold_presence diff_means sum_variance
#>  1: cg24192660            0.8981027            10  0.6911130   0.13719117
#>  2: cg03369247            0.9306799            10 -0.6070865   0.06291659
#>  3: cg11029367            0.8078551            10 -0.6247012   0.09282683
#>  4: cg14021880            0.9516236            10 -0.5730696   0.06488891
#>  5: cg06932535            0.6990152            10 -0.5998086   0.07776802
#>  6: cg16427420            0.7366863            10 -0.5869993   0.07927203
#>  7: cg04232282            0.7717972            10 -0.5580492   0.09075907
#>  8: cg15324723            0.8772958            10 -0.5215541   0.06901091
#>  9: cg02023138            0.8097305             6 -0.5213741   0.07679563
#> 10: cg17786959            0.8962509            10 -0.4598171   0.04412714
#> 11: cg26461417            0.8862681             4  0.4642382   0.04979224
#>     pred_type    cpg_score train_rank test_aupr
#>  1:      TRUE 8.981027e-01          1 0.6379077
#>  2:     FALSE 5.926112e-01          2 0.8777434
#>  3:     FALSE 5.759198e-01          3 0.9242849
#>  4:     FALSE 4.660049e-01          4 1.0000000
#>  5:     FALSE 4.231055e-01          5 0.8938304
#>  6:     FALSE 4.051134e-01          6 0.7787303
#>  7:     FALSE 3.278291e-01          7 0.7178212
#>  8:     FALSE 2.342303e-01          8 0.8107477
#>  9:     FALSE 7.186803e-02          9 0.3979287
#> 10:     FALSE 8.962509e-05         10 0.7282899
#> 11:      TRUE 1.702745e-06         11 0.8065877
#> 
#> $CELL_TYPE_MSCORFIBRO$elapsed_time
#> Time difference of 5.20626 secs
#> 
#> 
#> $CELL_TYPE_NEURONS
#> $CELL_TYPE_NEURONS$train_res
#> $CELL_TYPE_NEURONS$train_res$fold_id
#> # A tibble: 4,090 × 3
#>      Row Data     Fold  
#>    <int> <chr>    <chr> 
#>  1     1 Analysis Fold01
#>  2     1 Analysis Fold02
#>  3     1 Analysis Fold03
#>  4     1 Analysis Fold04
#>  5     1 Analysis Fold05
#>  6     1 Analysis Fold06
#>  7     1 Analysis Fold08
#>  8     1 Analysis Fold09
#>  9     1 Analysis Fold10
#> 10     2 Analysis Fold01
#> # … with 4,080 more rows
#> 
#> $CELL_TYPE_NEURONS$train_res$train_summary
#>             id     stat_origin mean_aupr fold_presence
#>  1: cg01149264      train_aupr 0.5835505            10
#>  2: cg01149264 validation_aupr 0.7109628            10
#>  3: cg02124957      train_aupr 1.0000000            10
#>  4: cg02124957 validation_aupr 1.0000000            10
#>  5: cg09146232      train_aupr 0.5192523            10
#>  6: cg09146232 validation_aupr 0.6272640            10
#>  7: cg09194778      train_aupr 0.9322908            10
#>  8: cg09194778 validation_aupr 0.9450694            10
#>  9: cg13618776      train_aupr 0.9300238            10
#> 10: cg13618776 validation_aupr 0.9391802            10
#> 11: cg13700051      train_aupr 1.0000000            10
#> 12: cg13700051 validation_aupr 1.0000000            10
#> 13: cg14356362      train_aupr 0.9783164            10
#> 14: cg14356362 validation_aupr 0.9450694            10
#> 15: cg17008486      train_aupr 0.9151931            10
#> 16: cg17008486 validation_aupr 0.9797267            10
#> 17: cg24548498      train_aupr 1.0000000            10
#> 18: cg24548498 validation_aupr 1.0000000            10
#> 19: cg26682499      train_aupr 0.9306130            10
#> 20: cg26682499 validation_aupr 0.9247961            10
#> 
#> $CELL_TYPE_NEURONS$train_res$dt_dmsv
#>             id diff_means sum_variance pred_type
#>  1: cg01149264 -0.5660675  0.017254806     FALSE
#>  2: cg02124957 -0.5272931  0.008677887     FALSE
#>  3: cg09146232 -0.7328558  0.031004450     FALSE
#>  4: cg09194778 -0.5813408  0.018445826     FALSE
#>  5: cg13618776  0.6295117  0.023289560      TRUE
#>  6: cg13700051 -0.5790852  0.010987231     FALSE
#>  7: cg14356362 -0.6419124  0.010098182     FALSE
#>  8: cg17008486 -0.7786229  0.032249027     FALSE
#>  9: cg24548498  0.5781948  0.008344877      TRUE
#> 10: cg26682499  0.5478238  0.017360200      TRUE
#> 
#> $CELL_TYPE_NEURONS$train_res$train_results
#>             id     stat_origin mean_aupr fold_presence diff_means sum_variance
#>  1: cg17008486 validation_aupr 0.9797267            10 -0.7786229  0.032249027
#>  2: cg09146232 validation_aupr 0.6272640            10 -0.7328558  0.031004450
#>  3: cg14356362 validation_aupr 0.9450694            10 -0.6419124  0.010098182
#>  4: cg13618776 validation_aupr 0.9391802            10  0.6295117  0.023289560
#>  5: cg13700051 validation_aupr 1.0000000            10 -0.5790852  0.010987231
#>  6: cg09194778 validation_aupr 0.9450694            10 -0.5813408  0.018445826
#>  7: cg24548498 validation_aupr 1.0000000            10  0.5781948  0.008344877
#>  8: cg01149264 validation_aupr 0.7109628            10 -0.5660675  0.017254806
#>  9: cg26682499 validation_aupr 0.9247961            10  0.5478238  0.017360200
#> 10: cg02124957 validation_aupr 1.0000000            10 -0.5272931  0.008677887
#>     pred_type   cpg_score train_rank
#>  1:     FALSE 0.489912359          1
#>  2:     FALSE 0.256551024          2
#>  3:     FALSE 0.215547366          3
#>  4:      TRUE 0.191034424          4
#>  5:     FALSE 0.103086204          5
#>  6:     FALSE 0.101664456          6
#>  7:      TRUE 0.101314909          7
#>  8:     FALSE 0.054878224          8
#>  9:      TRUE 0.037818868          9
#> 10:     FALSE 0.000050005         10
#> 
#> 
#> $CELL_TYPE_NEURONS$test_perf
#>             id validation_mean_aupr fold_presence diff_means sum_variance
#>  1: cg17008486            0.9797267            10 -0.7786229  0.032249027
#>  2: cg09146232            0.6272640            10 -0.7328558  0.031004450
#>  3: cg14356362            0.9450694            10 -0.6419124  0.010098182
#>  4: cg13618776            0.9391802            10  0.6295117  0.023289560
#>  5: cg13700051            1.0000000            10 -0.5790852  0.010987231
#>  6: cg09194778            0.9450694            10 -0.5813408  0.018445826
#>  7: cg24548498            1.0000000            10  0.5781948  0.008344877
#>  8: cg01149264            0.7109628            10 -0.5660675  0.017254806
#>  9: cg26682499            0.9247961            10  0.5478238  0.017360200
#> 10: cg02124957            1.0000000            10 -0.5272931  0.008677887
#>     pred_type   cpg_score train_rank test_aupr
#>  1:     FALSE 0.489912359          1 0.7248325
#>  2:     FALSE 0.256551024          2 0.1663329
#>  3:     FALSE 0.215547366          3 1.0000000
#>  4:      TRUE 0.191034424          4 0.9194535
#>  5:     FALSE 0.103086204          5 0.4922118
#>  6:     FALSE 0.101664456          6 0.8458771
#>  7:      TRUE 0.101314909          7 0.9215675
#>  8:     FALSE 0.054878224          8 0.2562279
#>  9:      TRUE 0.037818868          9 0.9920899
#> 10:     FALSE 0.000050005         10 0.8796454
#> 
#> $CELL_TYPE_NEURONS$elapsed_time
#> Time difference of 0.692908 secs
#> 
#>