Find simple CpG (CimpleG) signatures.

Train a classification model using (CpGs) as features for the given target data.

CimpleG(
  train_data,
  train_targets = NULL,
  targets,
  test_data = NULL,
  test_targets = NULL,
  method = c("CimpleG", "CimpleG_parab", "CimpleG_unscaled", "brute_force", "oner",
    "logistic_reg", "decision_tree", "boost_tree", "mlp", "rand_forest", "null_model"),
  pred_type = c("both", "hypo", "hyper"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  rank_method = c("default_rank", "a_rank", "c_rank", "ac_rank"),
  k_folds = 10,
  grid_n = 10,
  param_p = 2,
  quantile_threshold = 0.005,
  train_only = FALSE,
  run_parallel = FALSE,
  save_dir = NULL,
  save_format = c("zstd", "lz4", "gzip", "bzip2", "xz", "nocomp"),
  verbose = 1
)

Arguments

train_data

Training dataset. A matrix (s x f) with methylation data (Beta values) that will be used to train/find the predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns.

train_targets

A data frame with the training target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `train_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`.

targets

A string specifying the name of the column in `train_targets` to be used for training. Can be a character vector if there are several columns in `train_targets` to be used for training. If this argument is a character vector, CimpleG will search for the best predictors for each target sequentially or in parallel depending on the value of `run_parallel`

test_data

Testing dataset. A matrix (s x f) with methylation data (Beta values) that will be used to test the performance of the found predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

test_targets

A data frame with the testing target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `test_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

method

A string specifying the method or type of machine learning model/algorithm to be used for training. These are divided in two main groups. * The simple models (classifiers that use a single feature), `CimpleG` (default), `brute_force`, `CimpleG_unscaled` or `oner`; * the complex models (classifiers that use several features), `logistic_reg`, `decision_tree`, `boost_tree`, `mlp` or `rand_forest`.

pred_type

A string specifying the type of predictor/CpG to be searched for during training. Only used for simple models. One of `both` (default), `hypo` or `hyper`. If `hypo`, only hypomethylated predictors will be considered. If `hyper`, only hypermethylated predictors will be considered.

engine

A string specifying the machine learning engine behind `method`. Only used for complex models. Currently not in use.

rank_method

A string specifying the ranking strategy to rank the features during training.

k_folds

An integer specifying the number of folds (K) to be used in training for the stratified K-fold cross-validation procedure.

grid_n

An integer specifying the number of hyperparameter combinations to train for.

param_p

An even number in `sigma / (delta^param_p)`. Tunes how much weight will be given to delta when doing feature selection. Default is 2.

quantile_threshold

A number between 0 and 1. Determines how many features will be kept. Default is 0.005.

train_only

A boolean, if TRUE, CimpleG will only train (find predictors) but not test them against a test dataset.

run_parallel

A boolean, if `FALSE`, the default, it will search for predictors for multiple targets sequentially. If `TRUE` it will search for predictors for multiple targets at the same time (parallel processing) in order to save in computational time. You need to set up `future::plan()` before running this function.

save_dir

If defined it will save the resulting model to the given directory. Default is NULL.

save_format

Only used if save_dir is not NULL. One of "zstd", "lz4", "gzip", "bzip2","xz", "nocomp". zstd is the best option, fast compression and loading times, low space usage.

verbose

How verbose you want CimpleG to be while it is running. At 0, no message is displayed, at 3 every message is displayed. Default is 1.

Value

A CimpleG object with the results per target class.

Examples

library("CimpleG")

# read data
data(train_data)
data(train_targets)
data(test_data)
data(test_targets)

# run CimpleG
cimpleg_result <- CimpleG(
  train_data = train_data,
  train_targets = train_targets,
  test_data = test_data,
  test_targets = test_targets,
  method = "CimpleG",
  targets = c("CELL_TYPE_MSCORFIBRO","CELL_TYPE_NEURONS")
)
#> Training for target 'CELL_TYPE_MSCORFIBRO' with 'CimpleG' has finished.: 3.793 sec elapsed
#> Training for target 'CELL_TYPE_NEURONS' with 'CimpleG' has finished.: 0.638 sec elapsed

# check results
cimpleg_result$results
#> $CELL_TYPE_MSCORFIBRO
#> $CELL_TYPE_MSCORFIBRO$train_res
#> $CELL_TYPE_MSCORFIBRO$train_res$fold_id
#> # A tibble: 4,090 × 3
#>      Row Data     Fold  
#>    <int> <chr>    <chr> 
#>  1     1 Analysis Fold01
#>  2     1 Analysis Fold02
#>  3     1 Analysis Fold03
#>  4     1 Analysis Fold04
#>  5     1 Analysis Fold05
#>  6     1 Analysis Fold06
#>  7     1 Analysis Fold07
#>  8     1 Analysis Fold08
#>  9     1 Analysis Fold10
#> 10     2 Analysis Fold01
#> # … with 4,080 more rows
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_summary
#>             id     stat_origin mean_aupr mean_var_a fold_presence
#>  1: cg03369247      train_aupr 0.9308520  0.1709140            10
#>  2: cg03369247 validation_aupr 0.9303182  0.1709140            10
#>  3: cg06932535      train_aupr 0.6628177  0.2151255             9
#>  4: cg06932535 validation_aupr 0.6719639  0.2151255             9
#>  5: cg14021880      train_aupr 0.9483885  0.1976776            10
#>  6: cg14021880 validation_aupr 0.9396314  0.1976776            10
#>  7: cg16427420      train_aupr 0.6977225  0.2197322             5
#>  8: cg16427420 validation_aupr 0.6159754  0.2197322             5
#>  9: cg17786959      train_aupr 0.8717540  0.2089200            10
#> 10: cg17786959 validation_aupr 0.8767780  0.2089200            10
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$dt_dmsv
#>            id diff_means sum_variance pred_type
#> 1: cg03369247 -0.6070865   0.06291659     FALSE
#> 2: cg06932535 -0.5998086   0.07776802     FALSE
#> 3: cg14021880 -0.5730696   0.06488891     FALSE
#> 4: cg16427420 -0.5869993   0.07927203     FALSE
#> 5: cg17786959 -0.4598171   0.04412714     FALSE
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_results
#>            id fold_presence diff_means sum_variance pred_type mean_var_a
#> 1: cg03369247            10 -0.6070865   0.06291659     FALSE  0.1709140
#> 2: cg14021880            10 -0.5730696   0.06488891     FALSE  0.1976776
#> 3: cg17786959            10 -0.4598171   0.04412714     FALSE  0.2089200
#> 4: cg06932535             9 -0.5998086   0.07776802     FALSE  0.2151255
#> 5: cg16427420             5 -0.5869993   0.07927203     FALSE  0.2197322
#>    train_aupr validation_aupr cpg_score train_rank
#> 1:  0.9308520       0.9303182  5.649457          1
#> 2:  0.9483885       0.9396314  5.409834          2
#> 3:  0.8717540       0.8767780  4.020024          3
#> 4:  0.6628177       0.6719639  3.602760          4
#> 5:  0.6977225       0.6159754  1.927849          5
#> 
#> 
#> $CELL_TYPE_MSCORFIBRO$test_perf
#>            id fold_presence diff_means sum_variance pred_type mean_var_a
#> 1: cg03369247            10 -0.6070865   0.06291659     FALSE  0.1709140
#> 2: cg14021880            10 -0.5730696   0.06488891     FALSE  0.1976776
#> 3: cg17786959            10 -0.4598171   0.04412714     FALSE  0.2089200
#> 4: cg06932535             9 -0.5998086   0.07776802     FALSE  0.2151255
#> 5: cg16427420             5 -0.5869993   0.07927203     FALSE  0.2197322
#>    train_mean_aupr validation_mean_aupr cpg_score train_rank test_aupr
#> 1:       0.9308520            0.9303182  5.649457          1 0.8777434
#> 2:       0.9483885            0.9396314  5.409834          2 1.0000000
#> 3:       0.8717540            0.8767780  4.020024          3 0.7282899
#> 4:       0.6628177            0.6719639  3.602760          4 0.8938304
#> 5:       0.6977225            0.6159754  1.927849          5 0.7787303
#> 
#> $CELL_TYPE_MSCORFIBRO$elapsed_time
#> Time difference of 3.863025 secs
#> 
#> 
#> $CELL_TYPE_NEURONS
#> $CELL_TYPE_NEURONS$train_res
#> $CELL_TYPE_NEURONS$train_res$fold_id
#> # A tibble: 4,090 × 3
#>      Row Data     Fold  
#>    <int> <chr>    <chr> 
#>  1     1 Analysis Fold01
#>  2     1 Analysis Fold02
#>  3     1 Analysis Fold03
#>  4     1 Analysis Fold04
#>  5     1 Analysis Fold05
#>  6     1 Analysis Fold06
#>  7     1 Analysis Fold08
#>  8     1 Analysis Fold09
#>  9     1 Analysis Fold10
#> 10     2 Analysis Fold01
#> # … with 4,080 more rows
#> 
#> $CELL_TYPE_NEURONS$train_res$train_summary
#>             id     stat_origin mean_aupr mean_var_a fold_presence
#>  1: cg02124957      train_aupr 1.0000000 0.03120659            10
#>  2: cg02124957 validation_aupr 1.0000000 0.03120659            10
#>  3: cg13700051      train_aupr 1.0000000 0.03274830            10
#>  4: cg13700051 validation_aupr 1.0000000 0.03274830            10
#>  5: cg14356362      train_aupr 0.9784810 0.02450270            10
#>  6: cg14356362 validation_aupr 0.9306853 0.02450270            10
#>  7: cg21637776      train_aupr 0.9696044 0.04724633             8
#>  8: cg21637776 validation_aupr 0.9313367 0.04724633             8
#>  9: cg24548498      train_aupr 1.0000000 0.02494640            10
#> 10: cg24548498 validation_aupr 1.0000000 0.02494640            10
#> 
#> $CELL_TYPE_NEURONS$train_res$dt_dmsv
#>            id diff_means sum_variance pred_type
#> 1: cg02124957 -0.5272931  0.008677887     FALSE
#> 2: cg13700051 -0.5790852  0.010987231     FALSE
#> 3: cg14356362 -0.6419124  0.010098182     FALSE
#> 4: cg21637776 -0.3788893  0.006784796     FALSE
#> 5: cg24548498  0.5781948  0.008344877      TRUE
#> 
#> $CELL_TYPE_NEURONS$train_res$train_results
#>            id fold_presence diff_means sum_variance pred_type mean_var_a
#> 1: cg14356362            10 -0.6419124  0.010098182     FALSE 0.02450270
#> 2: cg13700051            10 -0.5790852  0.010987231     FALSE 0.03274830
#> 3: cg24548498            10  0.5781948  0.008344877      TRUE 0.02494640
#> 4: cg02124957            10 -0.5272931  0.008677887     FALSE 0.03120659
#> 5: cg21637776             8 -0.3788893  0.006784796     FALSE 0.04724633
#>    train_aupr validation_aupr cpg_score train_rank
#> 1:  0.9784810       0.9306853  6.127587          1
#> 2:  1.0000000       1.0000000  5.790852          2
#> 3:  1.0000000       1.0000000  5.781948          3
#> 4:  1.0000000       1.0000000  5.272931          4
#> 5:  0.9696044       0.9313367  2.880985          5
#> 
#> 
#> $CELL_TYPE_NEURONS$test_perf
#>            id fold_presence diff_means sum_variance pred_type mean_var_a
#> 1: cg14356362            10 -0.6419124  0.010098182     FALSE 0.02450270
#> 2: cg13700051            10 -0.5790852  0.010987231     FALSE 0.03274830
#> 3: cg24548498            10  0.5781948  0.008344877      TRUE 0.02494640
#> 4: cg02124957            10 -0.5272931  0.008677887     FALSE 0.03120659
#> 5: cg21637776             8 -0.3788893  0.006784796     FALSE 0.04724633
#>    train_mean_aupr validation_mean_aupr cpg_score train_rank test_aupr
#> 1:       0.9784810            0.9306853  6.127587          1 1.0000000
#> 2:       1.0000000            1.0000000  5.790852          2 0.4922118
#> 3:       1.0000000            1.0000000  5.781948          3 0.9215675
#> 4:       1.0000000            1.0000000  5.272931          4 0.8796454
#> 5:       0.9696044            0.9313367  2.880985          5 0.6205813
#> 
#> $CELL_TYPE_NEURONS$elapsed_time
#> Time difference of 0.6846859 secs
#> 
#>