Find simple CpG (CimpleG) signatures.

Train a classification model using (CpGs) as features for the given target data.

CimpleG(
  train_data,
  train_targets = NULL,
  targets,
  test_data = NULL,
  test_targets = NULL,
  method = c("CimpleG", "CimpleG_parab", "CimpleG_unscaled", "brute_force", "oner",
    "logistic_reg", "decision_tree", "boost_tree", "mlp", "rand_forest", "null_model"),
  pred_type = c("both", "hypo", "hyper"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  k_folds = 10,
  grid_n = 10,
  param_p = 4,
  quantile_threshold = 0.01,
  train_only = FALSE,
  run_parallel = FALSE,
  save_dir = NULL,
  save_format = c("zstd", "lz4", "gzip", "bzip2", "xz", "nocomp"),
  verbose = 1
)

Arguments

train_data

Training dataset. A matrix (s x f) with methylation data (Beta values) that will be used to train/find the predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns.

train_targets

A data frame with the training target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `train_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`.

targets

A string specifying the name of the column in `train_targets` to be used for training. Can be a character vector if there are several columns in `train_targets` to be used for training. If this argument is a character vector, CimpleG will search for the best predictors for each target sequentially or in parallel depending on the value of `run_parallel`

test_data

Testing dataset. A matrix (s x f) with methylation data (Beta values) that will be used to test the performance of the found predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

test_targets

A data frame with the testing target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `test_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

method

A string specifying the method or type of machine learning model/algorithm to be used for training. These are divided in two main groups. * The simple models (classifiers that use a single feature), `CimpleG` (default), `brute_force`, `CimpleG_unscaled` or `oner`; * the complex models (classifiers that use several features), `logistic_reg`, `decision_tree`, `boost_tree`, `mlp` or `rand_forest`.

pred_type

A string specifying the type of predictor/CpG to be searched for during training. Only used for simple models. One of `both` (default), `hypo` or `hyper`. If `hypo`, only hypomethylated predictors will be considered. If `hyper`, only hypermethylated predictors will be considered.

engine

A string specifying the machine learning engine behind `method`. Only used for complex models. Currently not in use.

k_folds

An integer specifying the number of folds (K) to be used in training for the stratified K-fold cross-validation procedure.

grid_n

An integer specifying the number of hyperparameter combinations to train for.

param_p

An even number in `sigma / (delta^param_p)`. Tunes how much weight will be given to delta when doing feature selection. Default is 4.

quantile_threshold

A number between 0 and 1. Determines how many features will be kept. Default is 0.001.

train_only

A boolean, if TRUE, CimpleG will only train (find predictors) but not test them against a test dataset.

run_parallel

A boolean, if `FALSE`, the default, it will search for predictors for multiple targets sequentially. If `TRUE` it will search for predictors for multiple targets at the same time (parallel processing) in order to save in computational time. You need to set up `future::plan()` before running this function.

save_dir

If defined it will save the resulting model to the given directory. Default is NULL.

save_format

Only used if save_dir is not NULL. One of "zstd", "lz4", "gzip", "bzip2","xz", "nocomp". zstd is the best option, fast compression and loading times, low space usage.

verbose

How verbose you want CimpleG to be while it is running. At 0, no message is displayed, at 3 every message is displayed. Default is 1.

Value

A CimpleG object with the results per target class.

Examples

library("CimpleG")

# read data
data(train_data)
data(train_targets)
data(test_data)
data(test_targets)

# run CimpleG
cimpleg_result <- CimpleG(
  train_data = train_data,
  train_targets = train_targets,
  test_data = test_data,
  test_targets = test_targets,
  method = "CimpleG",
  targets = c("CELL_TYPE_MSCORFIBRO","CELL_TYPE_NEURONS")
)
#> Training for target 'CELL_TYPE_MSCORFIBRO' with 'CimpleG' has finished.: 5.971 sec elapsed
#> Training for target 'CELL_TYPE_NEURONS' with 'CimpleG' has finished.: 1.372 sec elapsed

# check results
cimpleg_result$results
#> $CELL_TYPE_MSCORFIBRO
#> $CELL_TYPE_MSCORFIBRO$train_res
#> $CELL_TYPE_MSCORFIBRO$train_res$fold_id
#> # A tibble: 4,090 × 3
#>      Row Data     Fold  
#>    <int> <chr>    <chr> 
#>  1     1 Analysis Fold01
#>  2     1 Analysis Fold02
#>  3     1 Analysis Fold03
#>  4     1 Analysis Fold04
#>  5     1 Analysis Fold05
#>  6     1 Analysis Fold07
#>  7     1 Analysis Fold08
#>  8     1 Analysis Fold09
#>  9     1 Analysis Fold10
#> 10     2 Analysis Fold01
#> # … with 4,080 more rows
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_summary
#>             id     stat_origin mean_aupr fold_presence
#>  1: cg02023138      train_aupr 0.8268319             7
#>  2: cg02023138 validation_aupr 0.8570862             7
#>  3: cg03369247      train_aupr 0.9308630            10
#>  4: cg03369247 validation_aupr 0.9337021            10
#>  5: cg04232282      train_aupr 0.7346665            10
#>  6: cg04232282 validation_aupr 0.7669269            10
#>  7: cg06932535      train_aupr 0.6619573            10
#>  8: cg06932535 validation_aupr 0.7142441            10
#>  9: cg11029367      train_aupr 0.7912886            10
#> 10: cg11029367 validation_aupr 0.8130550            10
#> 11: cg14021880      train_aupr 0.9487070            10
#> 12: cg14021880 validation_aupr 0.9630081            10
#> 13: cg15324723      train_aupr 0.8775226            10
#> 14: cg15324723 validation_aupr 0.8825368            10
#> 15: cg16427420      train_aupr 0.6858212            10
#> 16: cg16427420 validation_aupr 0.7190267            10
#> 17: cg17786959      train_aupr 0.8718496            10
#> 18: cg17786959 validation_aupr 0.8924140            10
#> 19: cg24192660      train_aupr 0.9015798            10
#> 20: cg24192660 validation_aupr 0.8991319            10
#> 21: cg26461417      train_aupr 0.9399913             3
#> 22: cg26461417 validation_aupr 0.8428279             3
#>             id     stat_origin mean_aupr fold_presence
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$dt_dmsv
#>             id diff_means sum_variance pred_type   z_score
#>  1: cg02023138 -0.5213741   0.07679563     FALSE -3.552706
#>  2: cg03369247 -0.6070865   0.06291659     FALSE -4.178887
#>  3: cg04232282 -0.5580492   0.09075907     FALSE -3.820640
#>  4: cg06932535 -0.5998086   0.07776802     FALSE -4.125717
#>  5: cg11029367 -0.6247012   0.09282683     FALSE -4.307574
#>  6: cg14021880 -0.5730696   0.06488891     FALSE -3.930373
#>  7: cg15324723 -0.5215541   0.06901091     FALSE -3.554021
#>  8: cg16427420 -0.5869993   0.07927203     FALSE -4.032138
#>  9: cg17786959 -0.4598171   0.04412714     FALSE -3.102995
#> 10: cg24192660  0.6911130   0.13719117      TRUE  5.305248
#> 11: cg26461417  0.4642382   0.04979224      TRUE  3.647789
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_results
#>             id fold_presence   z_score diff_means sum_variance pred_type
#>  1: cg24192660            10  5.305248  0.6911130   0.13719117      TRUE
#>  2: cg03369247            10 -4.178887 -0.6070865   0.06291659     FALSE
#>  3: cg11029367            10 -4.307574 -0.6247012   0.09282683     FALSE
#>  4: cg14021880            10 -3.930373 -0.5730696   0.06488891     FALSE
#>  5: cg06932535            10 -4.125717 -0.5998086   0.07776802     FALSE
#>  6: cg16427420            10 -4.032138 -0.5869993   0.07927203     FALSE
#>  7: cg04232282            10 -3.820640 -0.5580492   0.09075907     FALSE
#>  8: cg15324723            10 -3.554021 -0.5215541   0.06901091     FALSE
#>  9: cg02023138             7 -3.552706 -0.5213741   0.07679563     FALSE
#> 10: cg26461417             3  3.647789  0.4642382   0.04979224      TRUE
#> 11: cg17786959            10 -3.102995 -0.4598171   0.04412714     FALSE
#>     train_aupr validation_aupr    cpg_score cpg_z_score train_rank
#>  1:  0.9015798       0.8991319 0.8991318751   4.7701172          1
#>  2:  0.9308630       0.9337021 0.5945355948   3.9018359          2
#>  3:  0.7912886       0.8130550 0.5796267418   3.5022942          3
#>  4:  0.9487070       0.9630081 0.4715797838   3.7849813          4
#>  5:  0.6619573       0.7142441 0.4323233258   2.9467694          5
#>  6:  0.6858212       0.7190267 0.3954021912   2.8992149          6
#>  7:  0.7346665       0.7669269 0.3257603901   2.9301513          7
#>  8:  0.8775226       0.8825368 0.2356296252   3.1365544          8
#>  9:  0.8268319       0.8570862 0.1597173606   2.1314828          9
#> 10:  0.9399913       0.8428279 0.0048578558   0.9223376         10
#> 11:  0.8718496       0.8924140 0.0000892414   2.7691559         11
#> 
#> 
#> $CELL_TYPE_MSCORFIBRO$test_perf
#>             id fold_presence   z_score diff_means sum_variance pred_type
#>  1: cg24192660            10  5.305248  0.6911130   0.13719117      TRUE
#>  2: cg03369247            10 -4.178887 -0.6070865   0.06291659     FALSE
#>  3: cg11029367            10 -4.307574 -0.6247012   0.09282683     FALSE
#>  4: cg14021880            10 -3.930373 -0.5730696   0.06488891     FALSE
#>  5: cg06932535            10 -4.125717 -0.5998086   0.07776802     FALSE
#>  6: cg16427420            10 -4.032138 -0.5869993   0.07927203     FALSE
#>  7: cg04232282            10 -3.820640 -0.5580492   0.09075907     FALSE
#>  8: cg15324723            10 -3.554021 -0.5215541   0.06901091     FALSE
#>  9: cg02023138             7 -3.552706 -0.5213741   0.07679563     FALSE
#> 10: cg26461417             3  3.647789  0.4642382   0.04979224      TRUE
#> 11: cg17786959            10 -3.102995 -0.4598171   0.04412714     FALSE
#>     train_mean_aupr validation_mean_aupr    cpg_score cpg_z_score train_rank
#>  1:       0.9015798            0.8991319 0.8991318751   4.7701172          1
#>  2:       0.9308630            0.9337021 0.5945355948   3.9018359          2
#>  3:       0.7912886            0.8130550 0.5796267418   3.5022942          3
#>  4:       0.9487070            0.9630081 0.4715797838   3.7849813          4
#>  5:       0.6619573            0.7142441 0.4323233258   2.9467694          5
#>  6:       0.6858212            0.7190267 0.3954021912   2.8992149          6
#>  7:       0.7346665            0.7669269 0.3257603901   2.9301513          7
#>  8:       0.8775226            0.8825368 0.2356296252   3.1365544          8
#>  9:       0.8268319            0.8570862 0.1597173606   2.1314828          9
#> 10:       0.9399913            0.8428279 0.0048578558   0.9223376         10
#> 11:       0.8718496            0.8924140 0.0000892414   2.7691559         11
#>     test_aupr
#>  1: 0.6379077
#>  2: 0.8777434
#>  3: 0.9242849
#>  4: 1.0000000
#>  5: 0.8938304
#>  6: 0.7787303
#>  7: 0.7178212
#>  8: 0.8107477
#>  9: 0.3979287
#> 10: 0.8065877
#> 11: 0.7282899
#> 
#> $CELL_TYPE_MSCORFIBRO$elapsed_time
#> Time difference of 6.048412 secs
#> 
#> 
#> $CELL_TYPE_NEURONS
#> $CELL_TYPE_NEURONS$train_res
#> $CELL_TYPE_NEURONS$train_res$fold_id
#> # A tibble: 4,090 × 3
#>      Row Data     Fold  
#>    <int> <chr>    <chr> 
#>  1     1 Analysis Fold01
#>  2     1 Analysis Fold02
#>  3     1 Analysis Fold03
#>  4     1 Analysis Fold05
#>  5     1 Analysis Fold06
#>  6     1 Analysis Fold07
#>  7     1 Analysis Fold08
#>  8     1 Analysis Fold09
#>  9     1 Analysis Fold10
#> 10     2 Analysis Fold01
#> # … with 4,080 more rows
#> 
#> $CELL_TYPE_NEURONS$train_res$train_summary
#>             id     stat_origin mean_aupr fold_presence
#>  1: cg01149264      train_aupr 0.5843366            10
#>  2: cg01149264 validation_aupr 0.7397310            10
#>  3: cg02124957      train_aupr 1.0000000            10
#>  4: cg02124957 validation_aupr 1.0000000            10
#>  5: cg09146232      train_aupr 0.5201077            10
#>  6: cg09146232 validation_aupr 0.6874062            10
#>  7: cg09194778      train_aupr 0.9322650            10
#>  8: cg09194778 validation_aupr 0.9247961            10
#>  9: cg13618776      train_aupr 0.9309797            10
#> 10: cg13618776 validation_aupr 0.9797267            10
#> 11: cg13700051      train_aupr 1.0000000            10
#> 12: cg13700051 validation_aupr 1.0000000            10
#> 13: cg14356362      train_aupr 0.9793310            10
#> 14: cg14356362 validation_aupr 1.0000000            10
#> 15: cg17008486      train_aupr 0.9145727            10
#> 16: cg17008486 validation_aupr 0.9797267            10
#> 17: cg23859313      train_aupr 0.5839710             1
#> 18: cg23859313 validation_aupr 0.2337616             1
#> 19: cg24548498      train_aupr 1.0000000            10
#> 20: cg24548498 validation_aupr 1.0000000            10
#> 21: cg26682499      train_aupr 0.9314835             9
#> 22: cg26682499 validation_aupr 0.9549483             9
#>             id     stat_origin mean_aupr fold_presence
#> 
#> $CELL_TYPE_NEURONS$train_res$dt_dmsv
#>             id diff_means sum_variance pred_type   z_score
#>  1: cg01149264 -0.5660675  0.017254806     FALSE -3.571634
#>  2: cg02124957 -0.5272931  0.008677887     FALSE -3.327860
#>  3: cg09146232 -0.7328558  0.031004450     FALSE -4.620228
#>  4: cg09194778 -0.5813408  0.018445826     FALSE -3.667657
#>  5: cg13618776  0.6295117  0.023289560      TRUE  3.944941
#>  6: cg13700051 -0.5790852  0.010987231     FALSE -3.653476
#>  7: cg14356362 -0.6419124  0.010098182     FALSE -4.048469
#>  8: cg17008486 -0.7786229  0.032249027     FALSE -4.907965
#>  9: cg23859313 -0.5748745  0.026046850     FALSE -3.627004
#> 10: cg24548498  0.5781948  0.008344877      TRUE  3.622313
#> 11: cg26682499  0.5478238  0.017360200      TRUE  3.431371
#> 
#> $CELL_TYPE_NEURONS$train_res$train_results
#>             id fold_presence   z_score diff_means sum_variance pred_type
#>  1: cg17008486            10 -4.907965 -0.7786229  0.032249027     FALSE
#>  2: cg09146232            10 -4.620228 -0.7328558  0.031004450     FALSE
#>  3: cg14356362            10 -4.048469 -0.6419124  0.010098182     FALSE
#>  4: cg13618776            10  3.944941  0.6295117  0.023289560      TRUE
#>  5: cg13700051            10 -3.653476 -0.5790852  0.010987231     FALSE
#>  6: cg24548498            10  3.622313  0.5781948  0.008344877      TRUE
#>  7: cg09194778            10 -3.667657 -0.5813408  0.018445826     FALSE
#>  8: cg01149264            10 -3.571634 -0.5660675  0.017254806     FALSE
#>  9: cg26682499             9  3.431371  0.5478238  0.017360200      TRUE
#> 10: cg23859313             1 -3.627004 -0.5748745  0.026046850     FALSE
#> 11: cg02124957            10 -3.327860 -0.5272931  0.008677887     FALSE
#>     train_aupr validation_aupr   cpg_score cpg_z_score train_rank
#>  1:  0.9145727       0.9797267 0.979726745  4.80846420          1
#>  2:  0.5201077       0.6874062 0.562242165  3.17597337          2
#>  3:  0.9793310       1.0000000 0.456105798  4.04846917          3
#>  4:  0.9309797       0.9797267 0.398523763  3.86496393          4
#>  5:  1.0000000       1.0000000 0.206151793  3.65347630          5
#>  6:  1.0000000       1.0000000 0.202609557  3.62231318          6
#>  7:  0.9322650       0.9247961 0.198947286  3.39183519          7
#>  8:  0.5843366       0.7397310 0.114186184  2.64204885          8
#>  9:  0.9314835       0.9549483 0.070286428  2.94910411          9
#> 10:  0.5839710       0.2337616 0.004427439  0.08478541         10
#> 11:  1.0000000       1.0000000 0.000100000  3.32786044         11
#> 
#> 
#> $CELL_TYPE_NEURONS$test_perf
#>             id fold_presence   z_score diff_means sum_variance pred_type
#>  1: cg17008486            10 -4.907965 -0.7786229  0.032249027     FALSE
#>  2: cg09146232            10 -4.620228 -0.7328558  0.031004450     FALSE
#>  3: cg14356362            10 -4.048469 -0.6419124  0.010098182     FALSE
#>  4: cg13618776            10  3.944941  0.6295117  0.023289560      TRUE
#>  5: cg13700051            10 -3.653476 -0.5790852  0.010987231     FALSE
#>  6: cg24548498            10  3.622313  0.5781948  0.008344877      TRUE
#>  7: cg09194778            10 -3.667657 -0.5813408  0.018445826     FALSE
#>  8: cg01149264            10 -3.571634 -0.5660675  0.017254806     FALSE
#>  9: cg26682499             9  3.431371  0.5478238  0.017360200      TRUE
#> 10: cg23859313             1 -3.627004 -0.5748745  0.026046850     FALSE
#> 11: cg02124957            10 -3.327860 -0.5272931  0.008677887     FALSE
#>     train_mean_aupr validation_mean_aupr   cpg_score cpg_z_score train_rank
#>  1:       0.9145727            0.9797267 0.979726745  4.80846420          1
#>  2:       0.5201077            0.6874062 0.562242165  3.17597337          2
#>  3:       0.9793310            1.0000000 0.456105798  4.04846917          3
#>  4:       0.9309797            0.9797267 0.398523763  3.86496393          4
#>  5:       1.0000000            1.0000000 0.206151793  3.65347630          5
#>  6:       1.0000000            1.0000000 0.202609557  3.62231318          6
#>  7:       0.9322650            0.9247961 0.198947286  3.39183519          7
#>  8:       0.5843366            0.7397310 0.114186184  2.64204885          8
#>  9:       0.9314835            0.9549483 0.070286428  2.94910411          9
#> 10:       0.5839710            0.2337616 0.004427439  0.08478541         10
#> 11:       1.0000000            1.0000000 0.000100000  3.32786044         11
#>     test_aupr
#>  1: 0.7248325
#>  2: 0.1663329
#>  3: 1.0000000
#>  4: 0.9194535
#>  5: 0.4922118
#>  6: 0.9215675
#>  7: 0.8458771
#>  8: 0.2562279
#>  9: 0.9920899
#> 10: 0.2659656
#> 11: 0.8796454
#> 
#> $CELL_TYPE_NEURONS$elapsed_time
#> Time difference of 1.431276 secs
#> 
#>