Train a classification model using (CpGs) as features for the given target data.

Train a classification model using (CpGs) as features for the given target data.

CimpleG(
  train_data,
  train_targets,
  targets,
  test_data = NULL,
  test_targets = NULL,
  method = c("parab_scale", "adhoc", "parab", "oner", "logistic_reg", "decision_tree",
    "boost_tree", "mlp", "rand_forest", "null_model"),
  pred_type = c("both", "hypo", "hyper"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  k_folds = 10,
  grid_n = 10,
  n_repeats = 1,
  run_parallel = FALSE
)

Arguments

train_data
Training dataset. A matrix (s x f) with methylation data (Beta values) that will be used to train/find the predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns.

train_targets
A data frame with the training target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `train_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`.

targets
A string specifying the name of the column in `train_targets` to be used for training. Can be a character vector if there are several columns in `train_targets` to be used for training. If this argument is a character vector, CimpleG will search for the best predictors for each target sequentially or in parallel depending on the value of `run_parallel`

test_data
Testing dataset. A matrix (s x f) with methylation data (Beta values) that will be used to test the performance of the found predictors. Samples (s) must be in rows while features/CpGs (f) must be in columns. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

test_targets
A data frame with the testing target samples one-hot encoded. A data frame with at least 1 column, with as many rows and in the same order as `test_data`. Target columms need to be one-hot encoded, meaning that, for that column the target samples should be encoded as `1` while every other sample should be encoded as `0`. If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a stratified test dataset based on `train_targets` by removing 25 samples from `train_data` and `train_targets`.

method
A string specifying the method or type of machine learning model/algorithm to be used for training. These are divided in two main groups. * The simple models (classifiers that use a single feature), `parab_scale` (default), `adhoc`, `parab` or `oner`; * the complex models (classifiers that use several features), `logistic_reg`, `decision_tree`, `boost_tree`, `mlp` or `rand_forest`.

pred_type
A string specifying the type of predictor/CpG to be searched for during training. Only used for simple models. One of `both` (default), `hypo` or `hyper`. If `hypo`, only hypomethylated predictors will be considered. If `hyper`, only hypermethylated predictors will be considered.

engine
A string specifying the machine learning engine behind `method`. Only used for complex models. Currently not in use.

k_folds
An integer specifying the number of folds (K) to be used in training for the stratified K-fold cross-validation procedure.

grid_n
An integer specifying the number of hyperparameter combinations to train for.

n_repeats
An integer specifying the number of repeats (N) to be used in training for the stratified N-repeats K-fold cross-validation procedure. Currently not in use.

run_parallel
A boolean, if `FALSE`, the default, it will search for predictors for multiple targets sequentially. If `TRUE` it will search for predictors for multiple targets at the same time (parallel processing) in order to save in computational time. You need to set up `future::plan()` before running this function.

Examples

library("CimpleG")

# read data
data(train_data)
data(train_targets)
data(test_data)
data(test_targets)

# run CimpleG
cimpleg_result <- CimpleG(
  train_data = train_data,
  train_targets = train_targets,
  test_data = test_data,
  test_targets = test_targets,
  method = "parab_scale",
  targets = c("CELL_TYPE_MSCORFIBRO","CELL_TYPE_NEURONS")
)

#> Training for target 'CELL_TYPE_MSCORFIBRO' with 'parab_scale is starting...

#> Fold parabola parameter: 0.9

#> Fold01: 0.417 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold02: 0.229 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold03: 0.242 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold04: 0.234 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold05: 0.252 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold06: 0.23 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold07: 0.251 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold08: 0.228 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold09: 0.23 sec elapsed

#> Fold parabola parameter: 0.9

#> Fold10: 0.277 sec elapsed
#> Training for target 'CELL_TYPE_MSCORFIBRO' with 'parab_scale(both)' has finished.: 5.82 sec elapsed

#> Training final model...

#> Evaluating test data...

#> Training for target 'CELL_TYPE_NEURONS' with 'parab_scale is starting...

#> Fold parabola parameter: 0.4

#> Fold01: 0.135 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold02: 0.135 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold03: 0.124 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold04: 0.178 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold05: 0.131 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold06: 0.138 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold07: 0.138 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold08: 0.152 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold09: 0.174 sec elapsed

#> Fold parabola parameter: 0.4

#> Fold10: 0.138 sec elapsed
#> Training for target 'CELL_TYPE_NEURONS' with 'parab_scale(both)' has finished.: 1.462 sec elapsed

#> Training final model...

#> Evaluating test data...

# check results
cimpleg_result$results

#> $CELL_TYPE_MSCORFIBRO
#> $CELL_TYPE_MSCORFIBRO$train_res
#> $CELL_TYPE_MSCORFIBRO$train_res$CpG
#> [1] "cg24192660"
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$model
#> # A tibble: 1 x 7
#>   predictor  type  .metric mean_fold_performa…     n mean_parab_param diff_means
#>   <chr>      <chr> <chr>                 <dbl> <int>            <dbl>      <dbl>
#> 1 cg24192660 hyper pr_auc                0.907    10              0.9      0.691
#> 
#> $CELL_TYPE_MSCORFIBRO$train_res$train_summary
#> # A tibble: 432 x 7
#>    predictor  type  .metric mean_fold_perform…     n mean_parab_param diff_means
#>    <chr>      <chr> <chr>                <dbl> <int>            <dbl>      <dbl>
#>  1 cg24192660 hyper roc_auc              0.955    10              0.9      0.691
#>  2 cg24192660 hyper pr_auc               0.907    10              0.9      0.691
#>  3 cg03369247 hypo  roc_auc              0.861    10              0.9     -0.607
#>  4 cg11029367 hypo  roc_auc              0.853    10              0.9     -0.625
#>  5 cg03369247 hypo  pr_auc               0.827    10              0.9     -0.607
#>  6 cg14021880 hypo  roc_auc              0.817    10              0.9     -0.573
#>  7 cg06932535 hypo  roc_auc              0.799    10              0.9     -0.600
#>  8 cg14021880 hypo  pr_auc               0.793    10              0.9     -0.573
#>  9 cg16427420 hypo  roc_auc              0.787    10              0.9     -0.587
#> 10 cg04232282 hypo  roc_auc              0.748    10              0.9     -0.558
#> # … with 422 more rows
#> 
#> 
#> $CELL_TYPE_MSCORFIBRO$test_perf
#>        method  predictor  accuracy    f1      aupr
#> 1 parab_scale cg24192660 0.7823529 0.704 0.6378487
#> 
#> 
#> $CELL_TYPE_NEURONS
#> $CELL_TYPE_NEURONS$train_res
#> $CELL_TYPE_NEURONS$train_res$CpG
#> [1] "cg17008486"
#> 
#> $CELL_TYPE_NEURONS$train_res$model
#> # A tibble: 1 x 7
#>   predictor  type  .metric mean_fold_performa…     n mean_parab_param diff_means
#>   <chr>      <chr> <chr>                 <dbl> <int>            <dbl>      <dbl>
#> 1 cg17008486 hypo  pr_auc                0.942    10              0.4     -0.779
#> 
#> $CELL_TYPE_NEURONS$train_res$train_summary
#> # A tibble: 168 x 7
#>    predictor  type  .metric mean_fold_perform…     n mean_parab_param diff_means
#>    <chr>      <chr> <chr>                <dbl> <int>            <dbl>      <dbl>
#>  1 cg17008486 hypo  roc_auc              0.997    10              0.4     -0.779
#>  2 cg17008486 hypo  pr_auc               0.942    10              0.4     -0.779
#>  3 cg09146232 hypo  roc_auc              0.922    10              0.4     -0.733
#>  4 cg14356362 hypo  roc_auc              0.821    10              0.4     -0.642
#>  5 cg14356362 hypo  pr_auc               0.805    10              0.4     -0.642
#>  6 cg13618776 hyper roc_auc              0.804    10              0.4      0.630
#>  7 cg08390172 hyper roc_auc              0.779    10              0.4      0.644
#>  8 cg13618776 hyper pr_auc               0.773    10              0.4      0.630
#>  9 cg13700051 hypo  pr_auc               0.741    10              0.4     -0.579
#> 10 cg13700051 hypo  roc_auc              0.741    10              0.4     -0.579
#> # … with 158 more rows
#> 
#> 
#> $CELL_TYPE_NEURONS$test_perf
#>        method  predictor  accuracy     f1      aupr
#> 1 parab_scale cg17008486 0.9411765 0.6875 0.7239611
#> 
#>